TSExternsGenerator: subclasses can't have static members with the same name, but different signatures
This is a weird restriction in TypeScript that other languages don't have (and JS would actually allow it). Seems that they decided to do it because the value of `this` can change, but we can't use `this` in statics in Haxe, so that doesn't really apply to us. Regardless, it should be possible to workaround by ommitting fields when extending a class, but I don't have time yet, so skip conflicts for now to allow Feathers UI to generate valid TS externs.

Relevant link: https://github.com/microsoft/TypeScript/issues/4628#issuecomment-1147905253